fix: 新版numpy废弃了np.float

+++ model/ns3gym/ns3gym/ns3env.py
@@ -116,9 +116,9 @@ class Ns3ZmqBridge(object):
             elif mtype == pb.UINT:
                 mtype = np.uint
             elif mtype == pb.DOUBLE:
-                mtype = np.float
+                mtype = np.float64
             else:
-                mtype = np.float
+                mtype = np.float32
 
             space = spaces.Box(low=low, high=high, shape=shape, dtype=mtype)
